+++ test/connection.js
@@ -1,0 +1,2 @@
+require('dotenv').load()
+require('./config').connect(process.env.MONGODB_URI_TESTING)