+++ CHANGELOG.md
@@ -17,3 +17,16 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ### Added
 - created basic server.js file, without passport or routes.
 
+## [0.2] - 2017-04-14
+### Added
+- added a mongo connection index.js file that returns a promise
+
+## [0.3] - 2017-04-14
+### Added
+- added server tests
+### BUGS // TO DO
+- need to create a testing database && easy way to drop & restart
+
+## [0.4] - 2017-04-14
+### Added
+- added separate testing database, connection files etc.